% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_change.R
\name{calculate_fc}
\alias{calculate_fc}
\title{calculate_fc}
\usage{
calculate_fc(
  object,
  by,
  control_group,
  case_group,
  mean_median = c("mean", "median")
)
}
\arguments{
\item{object}{tidymass-class object.}

\item{by}{one column of sample_info from object.}

\item{control_group}{A character vector, names from the by.}

\item{case_group}{A character vector, names from the by.}

\item{mean_median}{mean or median.}
}
\value{
object with fold change in variable_info.
}
\description{
Calculate fold change.
}
\examples{
library(massdataset)
library(ggplot2)
data("expression_data")
data("sample_info")
data("sample_info_note")
data("variable_info")
data("variable_info_note")

object =
create_tidymass_class(
expression_data = expression_data,
sample_info = sample_info,
variable_info = variable_info,
sample_info_note = sample_info_note,
variable_info_note = variable_info_note
)
object =
  calculate_fc(
  object = object,
  by = "class",
  control_group = c("Blank", "QC"),
  case_group = c("Subject"),
  mean_median = "mean"
)
head(extract_variable_info(object))
object =
  calculate_fc(
    object = object,
    by = "class",
    control_group = c("Blank", "QC"),
    case_group = c("Subject"),
    mean_median = "median"
  )
head(extract_variable_info(object))
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
