% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_group_test.R
\name{calculate_p_value}
\alias{calculate_p_value}
\title{calculate_p_value}
\usage{
calculate_p_value(
  object,
  by,
  control_group,
  case_group,
  method = c("t", "wilcox"),
  p_adjust_methods = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr",
    "none")
)
}
\arguments{
\item{object}{tidymass-class object.}

\item{by}{one column of sample_info from object.}

\item{control_group}{A character vector, names from the by.}

\item{case_group}{A character vector, names from the by.}

\item{method}{t test or wilcox test.}

\item{p_adjust_methods}{see ?p.adjust}
}
\value{
object with fold change in variable_info.
}
\description{
Calculate p values for variables.
}
\examples{
library(massdataset)
library(ggplot2)
data("expression_data")
data("sample_info")
data("sample_info_note")
data("variable_info")
data("variable_info_note")

object =
create_tidymass_class(
expression_data = expression_data,
sample_info = sample_info,
variable_info = variable_info,
sample_info_note = sample_info_note,
variable_info_note = variable_info_note
)

object =
  calculate_p_value(
  object = object,
  by = "class",
  control_group = c("Blank", "QC"),
  case_group = c("Subject"),
  method = "t",
  p_adjust_methods = "BH"
)
head(extract_variable_info(object))
object =
  calculate_p_value(
    object = object,
    by = "class",
    control_group = c("Blank", "QC"),
    case_group = c("Subject"),
    method = "wilcox",
    p_adjust_methods = "BH"
  )
head(extract_variable_info(object))
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
