% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_analysis.R, R/plsda_analysis.R
\name{convert_dummy_variable}
\alias{convert_dummy_variable}
\title{convert_dummy_variable}
\usage{
convert_dummy_variable(y)

convert_dummy_variable(y)
}
\arguments{
\item{y}{a vector of numeric or character}
}
\value{
ggplot2 object

ggplot2 object
}
\description{
convert_dummy_variable

convert_dummy_variable
}
\examples{
y = c(rep("a", 3), rep("b", 3))
convert_dummy_variable(y)
y = c(rep("a", 3), rep("b", 3))
convert_dummy_variable(y)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}

Xiaotao Shen
\email{shenxt1990@outlook.com}
}
