% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{Heatmap}
\alias{Heatmap}
\title{Heatmap}
\usage{
Heatmap(
  matrix,
  col,
  name = "value",
  na_col = "grey",
  color_space = "LAB",
  rect_gp = grid::gpar(col = NA),
  border = NA,
  border_gp = grid::gpar(col = "black"),
  cell_fun = NULL,
  layer_fun = NULL,
  jitter = FALSE,
  row_title = character(0),
  row_title_side = c("left", "right"),
  row_title_gp = grid::gpar(fontsize = 13.2),
  row_title_rot = switch(row_title_side[1], left = 90, right = 270),
  column_title = character(0),
  column_title_side = c("top", "bottom"),
  column_title_gp = grid::gpar(fontsize = 13.2),
  column_title_rot = 0,
  cluster_rows = TRUE,
  cluster_row_slices = TRUE,
  clustering_distance_rows = "euclidean",
  clustering_method_rows = "complete",
  row_dend_side = c("left", "right"),
  row_dend_width = unit(10, "mm"),
  show_row_dend = TRUE,
  row_dend_reorder = is.logical(cluster_rows) || is.function(cluster_rows),
  row_dend_gp = grid::gpar(),
  cluster_columns = TRUE,
  cluster_column_slices = TRUE,
  clustering_distance_columns = "euclidean",
  clustering_method_columns = "complete",
  column_dend_side = c("top", "bottom"),
  column_dend_height = unit(10, "mm"),
  show_column_dend = TRUE,
  column_dend_gp = grid::gpar(),
  column_dend_reorder = is.logical(cluster_columns) || is.function(cluster_columns),
  row_order = NULL,
  column_order = NULL,
  row_labels = rownames(matrix),
  row_names_side = c("right", "left"),
  show_row_names = TRUE,
  row_names_max_width = unit(6, "cm"),
  row_names_gp = grid::gpar(fontsize = 12),
  row_names_rot = 0,
  row_names_centered = FALSE,
  column_labels = colnames(matrix),
  column_names_side = c("bottom", "top"),
  show_column_names = TRUE,
  column_names_max_height = unit(6, "cm"),
  column_names_gp = grid::gpar(fontsize = 12),
  column_names_rot = 90,
  column_names_centered = FALSE,
  top_annotation = NULL,
  bottom_annotation = NULL,
  left_annotation = NULL,
  right_annotation = NULL,
  km = 1,
  split = NULL,
  row_km = km,
  row_km_repeats = 1,
  row_split = split,
  column_km = 1,
  column_km_repeats = 1,
  column_split = NULL,
  gap = unit(1, "mm"),
  row_gap = unit(1, "mm"),
  column_gap = unit(1, "mm"),
  show_parent_dend_line = ComplexHeatmap::ht_opt$show_parent_dend_line,
  heatmap_width = unit(1, "npc"),
  width = NULL,
  heatmap_height = unit(1, "npc"),
  height = NULL,
  show_heatmap_legend = TRUE,
  heatmap_legend_param = list(title = name),
  use_raster = NULL,
  raster_device = c("png", "jpeg", "tiff", "CairoPNG", "CairoJPEG", "CairoTIFF",
    "agg_png"),
  raster_quality = 1,
  raster_device_param = list(),
  raster_resize_mat = FALSE,
  raster_by_magick = requireNamespace("magick", quietly = TRUE),
  raster_magick_filter = NULL,
  post_fun = NULL
)
}
\arguments{
\item{matrix}{see ?ComplexHeatmap::Heatmap}

\item{col}{see ?ComplexHeatmap::Heatmap}

\item{name}{see ?ComplexHeatmap::Heatmap}

\item{na_col}{see ?ComplexHeatmap::Heatmap}

\item{color_space}{see ?ComplexHeatmap::Heatmap}

\item{rect_gp}{see ?ComplexHeatmap::Heatmap}

\item{border}{see ?ComplexHeatmap::Heatmap}

\item{border_gp}{see ?ComplexHeatmap::Heatmap}

\item{cell_fun}{see ?ComplexHeatmap::Heatmap}

\item{layer_fun}{see ?ComplexHeatmap::Heatmap}

\item{jitter}{see ?ComplexHeatmap::Heatmap}

\item{row_title}{see ?ComplexHeatmap::Heatmap}

\item{row_title_side}{see ?ComplexHeatmap::Heatmap}

\item{row_title_gp}{see ?ComplexHeatmap::Heatmap}

\item{row_title_rot}{see ?ComplexHeatmap::Heatmap}

\item{column_title}{see ?ComplexHeatmap::Heatmap}

\item{column_title_side}{see ?ComplexHeatmap::Heatmap}

\item{column_title_gp}{see ?ComplexHeatmap::Heatmap}

\item{column_title_rot}{see ?ComplexHeatmap::Heatmap}

\item{cluster_rows}{see ?ComplexHeatmap::Heatmap}

\item{cluster_row_slices}{see ?ComplexHeatmap::Heatmap}

\item{clustering_distance_rows}{see ?ComplexHeatmap::Heatmap}

\item{clustering_method_rows}{see ?ComplexHeatmap::Heatmap}

\item{row_dend_side}{see ?ComplexHeatmap::Heatmap}

\item{row_dend_width}{see ?ComplexHeatmap::Heatmap}

\item{show_row_dend}{see ?ComplexHeatmap::Heatmap}

\item{row_dend_reorder}{ee ?ComplexHeatmap::Heatmap}

\item{row_dend_gp}{see ?ComplexHeatmap::Heatmap}

\item{cluster_columns}{see ?ComplexHeatmap::Heatmap}

\item{cluster_column_slices}{see ?ComplexHeatmap::Heatmap}

\item{clustering_distance_columns}{see ?ComplexHeatmap::Heatmap}

\item{clustering_method_columns}{see ?ComplexHeatmap::Heatmap}

\item{column_dend_side}{see ?ComplexHeatmap::Heatmap}

\item{column_dend_height}{see ?ComplexHeatmap::Heatmap}

\item{show_column_dend}{see ?ComplexHeatmap::Heatmap}

\item{column_dend_gp}{see ?ComplexHeatmap::Heatmap}

\item{column_dend_reorder}{see ?ComplexHeatmap::Heatmap}

\item{row_order}{see ?ComplexHeatmap::Heatmap}

\item{column_order}{see ?ComplexHeatmap::Heatmap}

\item{row_labels}{see ?ComplexHeatmap::Heatmap}

\item{row_names_side}{see ?ComplexHeatmap::Heatmap}

\item{show_row_names}{see ?ComplexHeatmap::Heatmap}

\item{row_names_max_width}{see ?ComplexHeatmap::Heatmap}

\item{row_names_gp}{see ?ComplexHeatmap::Heatmap}

\item{row_names_rot}{see ?ComplexHeatmap::Heatmap}

\item{row_names_centered}{see ?ComplexHeatmap::Heatmap}

\item{column_labels}{see ?ComplexHeatmap::Heatmap}

\item{column_names_side}{see ?ComplexHeatmap::Heatmap}

\item{show_column_names}{see ?ComplexHeatmap::Heatmap}

\item{column_names_max_height}{see ?ComplexHeatmap::Heatmap}

\item{column_names_gp}{see ?ComplexHeatmap::Heatmap}

\item{column_names_rot}{see ?ComplexHeatmap::Heatmap}

\item{column_names_centered}{see ?ComplexHeatmap::Heatmap}

\item{top_annotation}{see ?ComplexHeatmap::Heatmap}

\item{bottom_annotation}{see ?ComplexHeatmap::Heatmap}

\item{left_annotation}{see ?ComplexHeatmap::Heatmap}

\item{right_annotation}{see ?ComplexHeatmap::Heatmap}

\item{km}{see ?ComplexHeatmap::Heatmap}

\item{split}{see ?ComplexHeatmap::Heatmap}

\item{row_km}{see ?ComplexHeatmap::Heatmap}

\item{row_km_repeats}{see ?ComplexHeatmap::Heatmap}

\item{row_split}{see ?ComplexHeatmap::Heatmap}

\item{column_km}{see ?ComplexHeatmap::Heatmap}

\item{column_km_repeats}{see ?ComplexHeatmap::Heatmap}

\item{column_split}{see ?ComplexHeatmap::Heatmap}

\item{gap}{see ?ComplexHeatmap::Heatmap}

\item{row_gap}{see ?ComplexHeatmap::Heatmap}

\item{column_gap}{see ?ComplexHeatmap::Heatmap}

\item{show_parent_dend_line}{see ?ComplexHeatmap::Heatmap}

\item{heatmap_width}{see ?ComplexHeatmap::Heatmap}

\item{width}{see ?ComplexHeatmap::Heatmap}

\item{heatmap_height}{see ?ComplexHeatmap::Heatmap}

\item{height}{see ?ComplexHeatmap::Heatmap}

\item{show_heatmap_legend}{see ?ComplexHeatmap::Heatmap}

\item{heatmap_legend_param}{see ?ComplexHeatmap::Heatmap}

\item{use_raster}{see ?ComplexHeatmap::Heatmap}

\item{raster_device}{see ?ComplexHeatmap::Heatmap}

\item{raster_quality}{see ?ComplexHeatmap::Heatmap}

\item{raster_device_param}{see ?ComplexHeatmap::Heatmap}

\item{raster_resize_mat}{see ?ComplexHeatmap::Heatmap}

\item{raster_by_magick}{see ?ComplexHeatmap::Heatmap}

\item{raster_magick_filter}{see ?ComplexHeatmap::Heatmap}

\item{post_fun}{see ?ComplexHeatmap::Heatmap}
}
\value{
A Heatmap-class object.
}
\description{
Heatmap function from ComplexHeatmap. Credit to Dr. Zuguang Gu.
}
\author{
Zuguang Gu
\email{z.gu@dkfz.de}
}
