% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_data.R
\docType{data}
\name{liver_aging_pos}
\alias{liver_aging_pos}
\title{liver aging dataset pos from MetDNA}
\format{
mass_dataset object
}
\usage{
liver_aging_pos
}
\description{
A mass_dataset object
A mass_dataset object
}
\keyword{datasets}
