% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_massstat_logo.R
\name{massstat_logo}
\alias{massstat_logo}
\title{massstat_logo}
\usage{
massstat_logo()
}
\value{
logo
}
\description{
massstat_logo
}
\examples{
massstat_logo()
}
\author{
Xiaotao Shen
\email{shenxt@stanford.edu}
}
