% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_utils.R
\name{massstat_packages}
\alias{massstat_packages}
\title{List all packages in the massstat}
\usage{
massstat_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include massstat in the list?}
}
\value{
massstat_packages
}
\description{
List all packages in the massstat
}
