% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_fc.R
\name{mutate_fc}
\alias{mutate_fc}
\title{mutate_fc}
\usage{
mutate_fc(
  object,
  control_sample_id,
  case_sample_id,
  mean_median = c("mean", "median")
)
}
\arguments{
\item{object}{tidymass-class object.}

\item{control_sample_id}{A character vector.}

\item{case_sample_id}{A character vector}

\item{mean_median}{mean or median.}
}
\value{
object with fold change (fc) in variable_info.
}
\description{
Calculate fold change.
}
\examples{
library(massdataset)
library(tidyverse)

data("liver_aging_pos")
liver_aging_pos

w_78 =
  liver_aging_pos \%>\%
  activate_mass_dataset(what = "sample_info") \%>\%
  dplyr::filter(group == "78W") \%>\%
  dplyr::pull(sample_id)

w_24 =
  liver_aging_pos \%>\%
  activate_mass_dataset(what = "sample_info") \%>\%
  dplyr::filter(group == "24W") \%>\%
  dplyr::pull(sample_id)


control_sample_id = w_24
case_sample_id = w_78

liver_aging_pos =
  mutate_fc(
    object = liver_aging_pos,
    control_sample_id = control_sample_id,
    case_sample_id = case_sample_id,
    mean_median = "mean"
  )

head(extract_variable_info(liver_aging_pos))

liver_aging_pos =
  mutate_fc(
    object = liver_aging_pos,
    control_sample_id = control_sample_id,
    case_sample_id = case_sample_id,
    mean_median = "median"
  )

head(extract_variable_info(liver_aging_pos))

extract_variable_info(liver_aging_pos) \%>\%
  ggplot(aes(fc, fc.1)) +
  geom_point()
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
