% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_group_test.R
\name{mutate_p_value}
\alias{mutate_p_value}
\title{mutate_p_value}
\usage{
mutate_p_value(
  object,
  by,
  control_group,
  case_group,
  method = c("t", "wilcox"),
  p_adjust_methods = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr",
    "none")
)
}
\arguments{
\item{object}{tidymass-class object.}

\item{by}{one column of sample_info from object.}

\item{control_group}{A character vector, names from the by.}

\item{case_group}{A character vector, names from the by.}

\item{method}{t test or wilcox test.}

\item{p_adjust_methods}{see ?p.adjust}
}
\value{
object with fold change in variable_info.
}
\description{
Calculate p values for variables.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
