% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_analysis.R
\name{pca_score_plot}
\alias{pca_score_plot}
\title{pca_score_plot}
\usage{
pca_score_plot(
  object,
  pca_object,
  color_by,
  point_alpha = 0.8,
  frame = TRUE,
  frame.type = "norm",
  ...
)
}
\arguments{
\item{object}{mass_dataset.}

\item{pca_object}{pca_object from run_pca()}

\item{color_by}{which column (sample_info) is used to color samples}

\item{point_alpha}{point_alpha}

\item{frame}{?ggplot2::autoplot}

\item{frame.type}{?ggplot2::autoplot}

\item{...}{other paramters for ggplot2::autoplot}
}
\value{
ggplot2 plot.
}
\description{
pca_score_plot
}
\examples{
library(massdataset)
library(demodata)
library(tidyverse)
data("liver_aging_pos", package = "demodata")
liver_aging_pos

pca_object =
liver_aging_pos \%>\%
  scale() \%>\%
  run_pca()

pca_score_plot(liver_aging_pos,
               pca_object,
               color_by = "group",
               loadings = TRUE) +
  ggsci::scale_fill_lancet() +
  ggsci::scale_color_lancet()

pca_score_plot(liver_aging_pos,
               pca_object,
               color_by = "group",
               frame = TRUE) +
  ggsci::scale_fill_lancet() +
  ggsci::scale_color_lancet() +
  ggrepel::geom_text_repel(aes(label = sample_id))
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
