% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda_analysis.R
\name{pls.mass_dataset}
\alias{pls.mass_dataset}
\title{plsda}
\usage{
\method{pls}{mass_dataset}(
  X,
  Y = NULL,
  ncomp = 2,
  scale = FALSE,
  mode = c("regression", "canonical", "invariant", "classic"),
  tol = 1e-06,
  max.iter = 100,
  near.zero.var = FALSE,
  logratio = "none",
  multilevel = NULL,
  all.outputs = TRUE
)
}
\arguments{
\item{X}{mass_dataset class}

\item{Y}{numeric matrix of response(s) with the rows as individual 
observations matching X. missing values (NAs) are allowed.}

\item{ncomp}{Positive Integer. The number of components to include in
the model. Default to 2.}

\item{scale}{Logical. If scale = TRUE, each block is standardized to zero
means and unit variances (default: TRUE)}

\item{mode}{Character string indicating the type of PLS algorithm to use.
One of "regression", "canonical", "invariant" or "classic". See Details.}

\item{tol}{Positive numeric used as convergence criteria/tolerance during
the iterative process. Default to 1e-06.}

\item{max.iter}{Integer, the maximum number of iterations. Default to 100.}

\item{near.zero.var}{Logical, see the internal nearZeroVar function
(should be set to TRUE in particular for data with many zero values).
Setting this argument to FALSE (when appropriate) will speed up the
computations. Default value is FALSE.}

\item{logratio}{Character, one of ('none','CLR') specifies the log ratio
transformation to deal with compositional values that may arise from
specific normalisation in sequencing data. Default to 'none'.
See ?logratio.transfo for details.}

\item{multilevel}{Numeric, design matrix for repeated measurement analysis,
where multilevel decomposition is required. For a one factor decomposition,
the repeated measures on each individual, i.e. the individuals ID is input
as the first column. For a 2 level factor decomposition then 2nd AND 3rd
columns indicate those factors. See examplesin ?spls.}

\item{all.outputs}{Logical. Computation can be faster when some specific
(and non-essential) outputs are not calculated. Default = TRUE.}
}
\value{
an object of class "pls"
}
\description{
plsda
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
