% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_analysis.R
\name{pls}
\alias{pls}
\title{pls}
\usage{
pls(
  X,
  Y = NULL,
  ncomp = 2,
  scale = FALSE,
  mode = c("regression", "canonical", "invariant", "classic"),
  tol = 1e-06,
  max.iter = 100,
  near.zero.var = FALSE,
  logratio = "none",
  multilevel = NULL,
  all.outputs = TRUE
)
}
\arguments{
\item{X}{mass_dataset}

\item{Y}{Y}

\item{ncomp}{ncomp}

\item{scale}{scale}

\item{mode}{mode}

\item{tol}{tol}

\item{max.iter}{max.iter}

\item{near.zero.var}{near.zero.var}

\item{logratio}{logratio}

\item{multilevel}{multilevel}

\item{all.outputs}{all.outputs}
}
\value{
A pls class
}
\description{
pls
}
