% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda_analysis.R
\name{plsda}
\alias{plsda}
\title{plsda from mixOmics, credit to mixOmics team}
\usage{
plsda(X, ...)
}
\arguments{
\item{X}{mass_dataset or matrix}

\item{...}{Other parameters}
}
\description{
plsda from mixOmics, credit to mixOmics team
}
