% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_method.R
\name{scale_data}
\alias{scale_data}
\title{scale_data}
\usage{
scale_data(
  x,
  center = TRUE,
  method = c("auto", "uv", "range", "pareto", "none")
)
}
\arguments{
\item{x}{mass_dataset}

\item{center}{center or not. TRUE or FALSE.}

\item{method}{auto, uv, range, pareto or none.}
}
\value{
scaled mass_dataset
}
\description{
Scale data.
}
\examples{
library(massdataset)
data("expression_data")
data("sample_info")
data("sample_info_note")
data("variable_info")
data("variable_info_note")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
    sample_info_note = sample_info_note,
    variable_info_note = variable_info_note
  )
object
scale_data(object, method = "auto")
scale_data(object, method = "range")
scale_data(object, method = "pareto")
scale_data(object, method = "none")
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
