% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_method.R
\name{scale_data.mass_dataset}
\alias{scale_data.mass_dataset}
\title{scale_data}
\usage{
\method{scale_data}{mass_dataset}(
  x,
  center = TRUE,
  method = c("auto", "uv", "range", "pareto", "none")
)
}
\arguments{
\item{x}{x}

\item{center}{center}

\item{method}{method for scaling.}
}
\value{
mass_dataset object
}
\description{
scale_data
}
