% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_cor.R
\name{cor_mass_dataset}
\alias{cor_mass_dataset}
\title{cor_mass_dataset}
\usage{
cor_mass_dataset(
  x,
  y = NULL,
  margin = c("variable", "sample"),
  use = c("everything", "all.obs", "complete.obs", "na.or.complete",
    "pairwise.complete.obs"),
  method = c("pearson", "kendall", "spearman")
)
}
\arguments{
\item{x}{mass_dataset class}

\item{y}{NULL}

\item{margin}{sample or variable}

\item{use}{an optional character string giving a method for computing
covariances in the presence of missing values. This must be
(an abbreviation of) one of the strings "everything", "all.obs",
"complete.obs", "na.or.complete", or "pairwise.complete.obs".}

\item{method}{a character string indicating which correlation coefficient
(or covariance) is to be computed. One of "pearson" (default),
"kendall", or "spearman": can be abbreviated.}
}
\value{
dist returns an object of class "dist".
}
\description{
cor_mass_dataset
}
\examples{
library(massdataset)
library(tidyverse)
data("expression_data")
data("sample_info")
data("sample_info_note")
data("variable_info")
data("variable_info_note")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
    sample_info_note = sample_info_note,
    variable_info_note = variable_info_note
  )
object
x =
  object \%>\%
  log(2) \%>\%
  scale()

variable_cor <-
  cor_mass_dataset(x = x, margin = "variable")
head(as.matrix(variable_cor)[, 1:5])
sample_cor <-
  cor_mass_dataset(x = x, margin = "sample")
head(as.matrix(sample_cor)[, 1:5])
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
