% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_p_value.R
\name{mutate_p_value}
\alias{mutate_p_value}
\title{mutate_p_value}
\usage{
mutate_p_value(
  object,
  control_sample_id,
  case_sample_id,
  method = c("t.test", "wilcox.test"),
  p_adjust_methods = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr",
    "none"),
  return_mass_dataset = TRUE,
  ...
)
}
\arguments{
\item{object}{tidymass-class object.}

\item{control_sample_id}{A character vector.}

\item{case_sample_id}{A character vector}

\item{method}{t test or wilcox test.}

\item{p_adjust_methods}{see ?p.adjust}

\item{return_mass_dataset}{logical default TRUE}

\item{...}{other parameters}
}
\value{
object with fold change in variable_info.
}
\description{
Calculate p values for variables.
}
\examples{
library(massdataset)
library(tidyverse)
data("liver_aging_pos")
liver_aging_pos

w_78 =
  liver_aging_pos \%>\%
  activate_mass_dataset(what = "sample_info") \%>\%
  dplyr::filter(group == "78W") \%>\%
  dplyr::pull(sample_id)

w_24 =
  liver_aging_pos \%>\%
  activate_mass_dataset(what = "sample_info") \%>\%
  dplyr::filter(group == "24W") \%>\%
  dplyr::pull(sample_id)

control_sample_id = w_24
case_sample_id = w_78

liver_aging_pos =
  mutate_p_value(
    object = liver_aging_pos,
    control_sample_id = control_sample_id,
    case_sample_id = case_sample_id,
    method = "t.test",
    p_adjust_methods = "BH"
  )

head(extract_variable_info(liver_aging_pos))

liver_aging_pos =
  mutate_p_value(
    object = liver_aging_pos,
    control_sample_id = control_sample_id,
    case_sample_id = case_sample_id,
    method = "wilcox.test",
    p_adjust_methods = "BH"
  )

head(extract_variable_info(liver_aging_pos))

extract_variable_info(liver_aging_pos) \%>\%
  ggplot(aes(-log(p_value_adjust, 10), -log(p_value_adjust.1, 10))) +
  geom_point()
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
