% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_analysis.R
\name{pca_score_plot}
\alias{pca_score_plot}
\title{pca_score_plot}
\usage{
pca_score_plot(
  object,
  pca_object,
  x = 1,
  y = 2,
  color_by,
  point_alpha = 1,
  point_size = 5,
  point_shape = 21,
  frame = TRUE,
  frame.type = "norm",
  ...
)
}
\arguments{
\item{object}{mass_dataset.}

\item{pca_object}{pca_object from run_pca()}

\item{x}{PC x for X axis, default is 1.}

\item{y}{PC y for Y axis, default is 2.}

\item{color_by}{which column (sample_info) is used to color samples}

\item{point_alpha}{point_alpha default is 1.}

\item{point_size}{point_size, default is 5.}

\item{point_shape}{point_shape default is 21.}

\item{frame}{?ggplot2::autoplot}

\item{frame.type}{?ggplot2::autoplot}

\item{...}{other paramters for ggplot2::autoplot}
}
\value{
ggplot2 plot.
}
\description{
pca_score_plot
}
\examples{
library(massdataset)
library(magrittr)
library(dplyr)
data("liver_aging_pos")
liver_aging_pos

pca_object =
  liver_aging_pos \%>\%
  scale() \%>\%
  run_pca()

pca_score_plot(liver_aging_pos,
               pca_object,
               color_by = "group",
               loadings = TRUE)

pca_score_plot(liver_aging_pos,
               pca_object,
               color_by = "group",
               frame = TRUE) +
  ggrepel::geom_text_repel(aes(label = sample_id))

pca_score_plot(liver_aging_pos,
               pca_object,
               x = 1,
               y = 3,
               color_by = "group",
               frame = TRUE) +
  ggrepel::geom_text_repel(aes(label = sample_id))
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
