% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_analysis.R, R/plsda_analysis.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{pls}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{mixOmics}{\code{\link[mixOmics]{pls}}, \code{\link[mixOmics]{pls}}}
}}

