% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_method.R
\name{scale_data}
\alias{scale_data}
\title{scale_data}
\usage{
scale_data(
  x,
  center = TRUE,
  method = c("auto", "uv", "range", "pareto", "none")
)
}
\arguments{
\item{x}{data.frame. Rows are variables and columns are samples.}

\item{center}{center or not. TRUE or FALSE.}

\item{method}{auto, uv, range, pareto or none.}
}
\value{
scaled data.frame.
}
\description{
Scale data.
}
\examples{
x <- as.data.frame(matrix(1:10, ncol = 5))
colnames(x) = letters[1:5]
scale_data(x, method = "auto")
t(scale(t(x)))
scale_data(x, method = "range")
scale_data(x, method = "pareto")
scale_data(x, method = "none")
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
