% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcono_plot2.R
\name{volcano_plot}
\alias{volcano_plot}
\alias{volcano_plot.mass_dataset}
\alias{volcano_plot.data.frame}
\title{volcano_plot}
\usage{
volcano_plot(
  object,
  fc_column_name = "fc",
  p_value_column_name = "p_value_adjust",
  labs_x = "log2(Fold change)",
  labs_y = "-log(p-adjust, 10)",
  fc_up_cutoff = 2,
  fc_down_cutoff = 0.5,
  p_value_cutoff = 0.05,
  line_color = "red",
  up_color = "#EE0000FF",
  down_color = "#3B4992FF",
  no_color = "#808180FF",
  point_size = 2,
  point_alpha = 1,
  point_size_scale = "log10_p",
  line_type = 1,
  add_text = FALSE,
  text_for = c("marker", "UP", "DOWM"),
  text_from = "variable_id"
)

\method{volcano_plot}{mass_dataset}(
  object,
  fc_column_name = "fc",
  p_value_column_name = "p_value_adjust",
  labs_x = "log2(Fold change)",
  labs_y = "-log(p-adjust, 10)",
  fc_up_cutoff = 2,
  fc_down_cutoff = 0.5,
  p_value_cutoff = 0.05,
  line_color = "red",
  up_color = "#EE0000FF",
  down_color = "#3B4992FF",
  no_color = "#808180FF",
  point_size = 2,
  point_alpha = 1,
  point_size_scale = "log10_p",
  line_type = 1,
  add_text = FALSE,
  text_for = c("marker", "UP", "DOWM"),
  text_from = "variable_id"
)

\method{volcano_plot}{data.frame}(
  object,
  fc_column_name = "fc",
  p_value_column_name = "p_value_adjust",
  labs_x = "log2(Fold change)",
  labs_y = "-log(p-adjust, 10)",
  fc_up_cutoff = 2,
  fc_down_cutoff = 0.5,
  p_value_cutoff = 0.05,
  line_color = "red",
  up_color = "#EE0000FF",
  down_color = "#3B4992FF",
  no_color = "#808180FF",
  point_size = 2,
  point_alpha = 1,
  point_size_scale = "log10_p",
  line_type = 1,
  add_text = FALSE,
  text_for = c("marker", "UP", "DOWM"),
  text_from = "variable_id"
)
}
\arguments{
\item{object}{mass_dataset object or data.frame}

\item{fc_column_name}{fc_column_name}

\item{p_value_column_name}{p_value_column_name}

\item{labs_x}{labs_x}

\item{labs_y}{labs_y}

\item{fc_up_cutoff}{fc_up_cutoff}

\item{fc_down_cutoff}{fc_down_cutoff}

\item{p_value_cutoff}{p_value_cutoff}

\item{line_color}{line_color}

\item{up_color}{up_color}

\item{down_color}{down_color}

\item{no_color}{no_color}

\item{point_size}{point_size}

\item{point_alpha}{point_alpha}

\item{point_size_scale}{point_size_scale}

\item{line_type}{line_type}

\item{add_text}{add_text}

\item{text_for}{text_for}

\item{text_from}{text_from}
}
\value{
ggplot2 object
}
\description{
Draw volcano plot.
}
\examples{
library(massdataset)
library(magrittr)
library(dplyr)

data("liver_aging_pos")
liver_aging_pos

w_78 =
  liver_aging_pos \%>\%
  activate_mass_dataset(what = "sample_info") \%>\%
  dplyr::filter(group == "78W") \%>\%
  dplyr::pull(sample_id)

w_24 =
  liver_aging_pos \%>\%
  activate_mass_dataset(what = "sample_info") \%>\%
  dplyr::filter(group == "24W") \%>\%
  dplyr::pull(sample_id)

control_sample_id = w_24
case_sample_id = w_78

liver_aging_pos =
  mutate_fc(
    object = liver_aging_pos,
    control_sample_id = control_sample_id,
    case_sample_id = case_sample_id,
    mean_median = "mean"
  )

liver_aging_pos =
  mutate_p_value(
    object = liver_aging_pos,
    control_sample_id = control_sample_id,
    case_sample_id = case_sample_id,
    method = "t.test",
    p_adjust_methods = "BH"
  )

object = liver_aging_pos
volcano_plot(
  object = object,
  fc_column_name = "fc",
  p_value_column_name = "p_value_adjust",
  labs_x = "log2(Fold change)",
  labs_y = "-log(p-adjust, 10)",
  fc_up_cutoff = 2,
  fc_down_cutoff = 0.5,
  p_value_cutoff = 0.05,
  add_text = TRUE
)
volcano_plot(
  object = object,
  fc_column_name = "fc",
  p_value_column_name = "p_value",
  labs_x = "log2(Fold change)",
  labs_y = "-log(p-value, 10)",
  fc_up_cutoff = 2,
  fc_down_cutoff = 0.5,
  p_value_cutoff = 0.05,
  add_text = FALSE,
  point_alpha = 0.5
)

volcano_plot(
  object = object,
  fc_column_name = "fc",
  p_value_column_name = "p_value",
  labs_x = "log2(Fold change)",
  labs_y = "-log(p-value, 10)",
  fc_up_cutoff = 2,
  fc_down_cutoff = 0.5,
  p_value_cutoff = 0.05,
  add_text = FALSE,
  point_alpha = 0.5,
  point_size_scale = "p_value"
) +
  scale_size_continuous(range = c(0.5, 3))
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
